requirements.txt
# ---- Core API ----
fastapi==0.128.0
uvicorn==0.40.0
starlette==0.50.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.22
anyio==4.12.1
h11==0.16.0
click==8.3.1

# ---- Audio Processing (Cross-platform) ----
numpy==2.3.5
librosa==0.10.2.post1
soundfile==0.13.1

# ---- Networking ----
requests==2.32.5
urllib3==2.6.3
charset-normalizer==3.4.4
idna==3.11
certifi==2026.1.4

# ---- ML Libraries ----
pandas==2.2.3
openpyxl==3.1.5
scikit-learn==1.6.1
joblib==1.4.2
datasets==3.2.0
huggingface_hub==0.27.1

# ---- Utilities ----
tqdm==4.67.2
typing_extensions==4.15.0
